Silence signed/unsigned comparison warnings in GCC.
Why do the wait APIs take a signed integer if they are going to be
compared to unsigned quantities?

+++ libraries/mbed/common/wait_api.c
@@ -26,5 +26,5 @@ void wait_ms(int ms) {
 
 void wait_us(int us) {
     uint32_t start = us_ticker_read();
-    while ((us_ticker_read() - start) < us);
+    while ((us_ticker_read() - start) < (uint32_t)us);
 }
